Added: (Common)   - HistoryInventoryEquipment class (DataBaseLocal)   - Method of returning the user by logic (Interface)   - Window for adding a comment for the history of equipment-inventory   - Custom modal window message box   - Custom TextBox controll   - Application icon   - Placeholder for images   - Possibility to leave a comment for equipment from the inventory   - Interface for the employee user (Logic)   - Possibility to leave a comment for equipment from the inventory   - Logic for the employee user   - Authorization logic
Deleted:
(Common)
  - HistoryAudienceEquipment class

Changed:
(Common)
  - Name of fields and class tables
(DataBaseLocal)
  - HistoryAudienceEquipment context has been replaced with HistoryInventoryEquipment
 - Local database reset
  - The request sent when checking the connection.
(DataBaseServer)
  - HistoryAudienceEquipment context has been replaced with HistoryInventoryEquipment
  - The request sent when checking the connection.
(Interface)
  - Redesigned item interface

Fixed:
(ValidationData)
  - Invalid error output
  - Checking for invalid cost field type

Other:
  - Various UI fixes, changes
  - Various Logic fixes, changes
  - Various bag fixes

+++ ValidationData/EquipmentValidator.cs
@@ -59,7 +59,7 @@ namespace ValidationData
         /// <returns>True if the cost is a valid integer, otherwise false.</returns>
         public static bool ValidateCost(string cost)
         {
-            if (!int.TryParse(cost, out _))
+            if (!decimal.TryParse(cost, out _))
             {
                 return false;
             }
